[c10d] Make reduce as a custom op
Summary:
This patch makes reduce as a custom op such that it's dispatcher
passable. It's one part of the effort to route comm ops to the dispatcher
such that tracing mechanisms that relies on the dispatcher can trace them,
e.g., LazyTensor and AOTAutograd.

Test Plan:
python test/distributed/test_c10d_nccl.py -k test_reduce_ops
python test/distributed/test_c10d_gloo.py -k test_reduce_basics
...and other existing distributed tests.

[ghstack-poisoned]

+++ torch/csrc/distributed/c10d/Types.hpp
@@ -34,8 +34,8 @@ struct AllreduceCoalescedOptions : AllreduceOptions {};
 
 struct ReduceOptions {
   ReduceOp reduceOp = ReduceOp::SUM;
-  int rootRank = 0;
-  int rootTensor = 0;
+  int64_t rootRank = 0;
+  int64_t rootTensor = 0;
   std::chrono::milliseconds timeout = kUnsetTimeout;
 };
